move mvcc_rec_header to mvcc.h
