+++ src/storage/btree_load.c
@@ -36,6 +36,7 @@
 #include "external_sort.h"
 #include "heap_file.h"
 #include "memory_private_allocator.hpp"
+#include "mvcc.h"
 #include "object_primitive.h"
 #include "object_representation_sr.h"
 #include "partition.h"
